Extract portfolio income activity contracts
